requirements.txt
bson==0.5.8
click==7.0
python_dateutil==2.8.0
pytz==2019.1
munch==2.3.2
werkzeug==0.15.4
python-snappy==0.5.4
confluent-kafka==1.0.1
PyMySQL==0.9.3
mysql-replication==0.19
pymongo==3.8.0
redis==3.2.1
celery==4.3.0
flower==0.9.3
docker==4.0.2
